Finished Storyboard Card Workout App
CardSelectionVC presents buttons to start and stop the shuffling of a deck of cards to determine the "workout" they will need to adhere to. When the shuffling starts the start button dynamically changes to a stop button to ensure the deck isn't continuously shuffling on a endless timer.

This is a storyboard implementation